polymarket_bot/requirements.txt
# Polymarket Trading Bot Requirements
# Pin versions to avoid compatibility issues

# Polymarket SDK (MUST pin these versions)
py-clob-client==0.34.5
web3==6.14.0  # Critical: Newer versions break py-clob-client

# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0

# Development/testing
pytest==7.4.4

# Install with:
# pip install -r requirements.txt